[zsh] do not start ssh-agent on graphical logins

+++ profile
@@ -59,8 +59,8 @@ fi
 # fi
 
 # Start ssh-agent if necessary.
-SSHAGENT=/usr/bin/ssh-agent
-if [ -z "$SSH_AUTH_SOCK" -a -x "$SSHAGENT" ]; then
+# SSHAGENT=/usr/bin/ssh-agent
+if [ -z "$DISPLAY" -a -z "$SSH_AUTH_SOCK" -a -x "$SSHAGENT" ]; then
     eval `$SSHAGENT -s`
     trap "kill $SSH_AGENT_PID" 0
 fi
